refactor: align naming
